pyproject.toml
[project]
name = "fishlog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "pillow>=10.4.0",
    "python-dotenv",
]
